"move devicepool to device_context"
